requirements.txt
paho-mqtt>=1.4.0
psutil>=5.6.3
pytz>=2019.2
Pillow>=6.1.0
pimoroni-bme280
pms5003
ltr559
st7735
ads1015